commit 3

+++ src/main/java/com/StudentLance/demo/DAO/CompanyDAO.java
@@ -9,7 +9,7 @@ import org.springframework.data.repository.query.Param;
 
 import java.util.List;
 @Repository
-public interface CompanyDAO extends JpaRepository<Company,Integer> {
+public interface CompanyDAO extends JpaRepository<Company,Long> {
 
     public Company findByCompanyId(int  companyId);
     public List<Company> findAll();
